requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==20.0.4
Jinja2==2.11.2
pipenv==2020.8.13
pycparser==2.20
pylint==2.5.3
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
Werkzeug==1.0.1
flask_mongoengine==1.0.0
python-decouple==3.1